Move categories component
