+++ src/App.js
@@ -3,7 +3,7 @@ import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import Navbar from './components/navbar';
 import Books from './components/bookReact';
-import Categories from './redux/categories/categories';
+import Categories from './components/categories';
 
 const App = () => (
   <>
